Fixes NNUE for FRC.
