+++ illumina/illumina.cpp
@@ -10,6 +10,7 @@ void init_attacks();
 void init_types();
 void init_zob();
 void init_search();
+void init_nnue();
 
 void init() {
     if (s_initialized) {
@@ -18,7 +19,7 @@ void init() {
 
     s_initialized = true;
 
-    init_eval();
+    init_nnue();
     init_types();
     init_zob();
     init_attacks();
